fix: replace FC by function / refacto code in creating participants-section.tsx file
